chat_completion/requirement.txt
python-dotenv
streamlit
mistralai